Moves the settings section to a scene
