+++ src/lib/Containers/index.tsx
@@ -1,10 +1,10 @@
 import Home from "../Scenes/Home"
+import MyProfile from "../Scenes/Settings/MyProfile"
 import Artist from "./Artist"
 import Conversation from "./Conversation"
 import Gene from "./Gene"
 import Inbox from "./Inbox"
 import Inquiry from "./Inquiry"
-import MyProfile from "./MyProfile"
 import Sale from "./Sale"
 import WorksForYou from "./WorksForYou"
