+++ storybook/storyLoader.js
@@ -24,8 +24,8 @@ function loadStories() {
   require('../src/lib/Containers/__stories__/Artist.story.tsx');
   require('../src/lib/Containers/__stories__/Gene.story.tsx');
   require('../src/lib/Containers/__stories__/Inquiry.story.tsx');
-  require('../src/lib/Containers/__stories__/MyProfile.story.tsx');
   require('../src/lib/Containers/__stories__/Sale.story.tsx');
+  require('../src/lib/Scenes/Settings/__stories__/MyProfile.story.tsx');
   
 }
